+++ NanoXLSX/Styles/NumberFormat.cs
@@ -1,9 +1,11 @@
 ﻿/*
  * NanoXLSX is a small .NET library to generate and read XLSX (Microsoft Excel 2007 or newer) files in an easy and native way  
- * Copyright Raphael Stoeckli © 2021
+ * Copyright Raphael Stoeckli © 2022
  * This library is licensed under the MIT License.
  * You find a copy of the license in project folder or on: http://opensource.org/licenses/MIT
  */
+
+using System.Collections.Generic;
 
 namespace NanoXLSX.Styles
 {
@@ -277,14 +279,12 @@ namespace NanoXLSX.Styles
         /// </returns>
         public override int GetHashCode()
         {
-            const int p = 251;
-            int r = 1;
-            r *= p + this.CustomFormatCode.GetHashCode();
-            r *= p + this.CustomFormatID;
-            r *= p + (int)this.Number;
-            return r;
+            int hashCode = 495605284;
+            hashCode = hashCode * -1521134295 + EqualityComparer<string>.Default.GetHashCode(CustomFormatCode);
+            hashCode = hashCode * -1521134295 + CustomFormatID.GetHashCode();
+            hashCode = hashCode * -1521134295 + Number.GetHashCode();
+            return hashCode;
         }
-
         #endregion
     }
 }