+++ NanoXLSX/Utils.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * NanoXLSX is a small .NET library to generate and read XLSX (Microsoft Excel 2007 or newer) files in an easy and native way
- * Copyright Raphael Stoeckli © 2021
+ * Copyright Raphael Stoeckli © 2022
  * This library is licensed under the MIT License.
  * You find a copy of the license in project folder or on: http://opensource.org/licenses/MIT
  */
@@ -9,7 +9,7 @@ using System;
 using System.Globalization;
 using System.Runtime.CompilerServices;
 using NanoXLSX.Exceptions;
-
+using FormatException = NanoXLSX.Exceptions.FormatException;
 
 namespace NanoXLSX
 {
@@ -138,7 +138,7 @@ namespace NanoXLSX
         /// The internal width deviates slightly from the column width, entered in Excel. Although internal, the default column width of 10 characters is visible in Excel as 10.71.
         /// The deviation depends on the maximum digit width of the default font, as well as its text padding and various constants.<br/>
         /// In case of the width 10.0 and the default digit width 7.0, as well as the padding 5.0 of the default font Calibri (size 11), 
-        /// the internal width is approximately 10.7142857 (rounded to 10.71).<br/> Note that the column hight is not affected by this consideration. 
+        /// the internal width is approximately 10.7142857 (rounded to 10.71).<br/> Note that the column height is not affected by this consideration. 
         /// The entered height in Excel is the actual height in the worksheet XML documents.<br/> 
         /// This method is derived from the Perl implementation by John McNamara (<a href="https://stackoverflow.com/a/5010899">https://stackoverflow.com/a/5010899</a>)<br/>
         /// See also: <a href="https://www.ecma-international.org/publications-and-standards/standards/ecma-376/">ECMA-376, Part 1, Chapter 18.3.1.13</a>
@@ -147,8 +147,13 @@ namespace NanoXLSX
         /// <param name="maxDigitWidth">Maximum digit with of the default font (default is 7.0 for Calibri, size 11)</param>
         /// <param name="textPadding">Text padding of the default font (default is 5.0 for Calibri, size 11)</param>
         /// <returns>The internal column width in characters, used in worksheet XML documents</returns>
+        /// <exception cref="FormatException">Throws a FormatException if the column width is out of range</exception>
         public static float GetInternalColumnWidth(float columnWidth, float maxDigitWidth = 7f, float textPadding = 5f)
         {
+            if (columnWidth < Worksheet.MIN_COLUMN_WIDTH || columnWidth > Worksheet.MAX_COLUMN_WIDTH)
+            {
+                throw new FormatException("The column width " + columnWidth + " is not valid. The valid range is between " + Worksheet.MIN_COLUMN_WIDTH + " and " + Worksheet.MAX_COLUMN_WIDTH);
+            }
             if (columnWidth <= 0f || maxDigitWidth <= 0f)
             {
                 return 0f;
@@ -171,9 +176,14 @@ namespace NanoXLSX
         /// Therefore, the originally defined row height will slightly deviate, based on this pixel snap</remarks>
         /// <param name="rowHeight">Target row height (displayed in Excel)</param>
         /// <returns>The internal row height which snaps to the nearest pixel</returns>
+        /// <exception cref="FormatException">Throws a FormatException if the row height is out of range</exception>
         public static float GetInternalRowHeight(float rowHeight)
         {
-            if (rowHeight <= 0f)
+            if (rowHeight < Worksheet.MIN_ROW_HEIGHT || rowHeight > Worksheet.MAX_ROW_HEIGHT)
+            {
+                throw new FormatException("The row height " + rowHeight + " is not valid. The valid range is between " + Worksheet.MIN_ROW_HEIGHT + " and " + Worksheet.MAX_ROW_HEIGHT);
+            }
+            if (rowHeight == 0f)
             {
                 return 0f;
             }
@@ -190,7 +200,7 @@ namespace NanoXLSX
         /// See also <see cref="GetInternalColumnWidth(float, float, float)"/> for additional details.<br/>
         /// This method is derived from the Perl implementation by John McNamara (<a href="https://stackoverflow.com/a/5010899">https://stackoverflow.com/a/5010899</a>)<br/>
         /// See also: <a href="https://www.ecma-international.org/publications-and-standards/standards/ecma-376/">ECMA-376, Part 1, Chapter 18.3.1.13</a><br/>
-        /// The three optional parameters maxDigitWidth and textPadding probably don't have to be changed ever.
+        /// The two optional parameters maxDigitWidth and textPadding probably don't have to be changed ever. Negative column widths are automatically transformed to 0.
         /// </remarks>
         /// <param name="width">Target column(s) width (one or more columns, displayed in Excel)</param>
         /// <param name="maxDigitWidth">Maximum digit with of the default font (default is 7.0 for Calibri, size 11)</param>
@@ -199,6 +209,11 @@ namespace NanoXLSX
         public static float GetInternalPaneSplitWidth(float width, float maxDigitWidth = 7f, float textPadding = 5f)
         {
             float pixels;
+            // TODO: Check the <1 part again. Leads always to 390
+            if (width <= 1f)
+            {
+                width = 0;
+            }
             if (width <= 1f)
             {
                 pixels = (float)Math.Floor(width / SPLIT_WIDTH_MULTIPLIER + SPLIT_WIDTH_OFFSET);
@@ -216,12 +231,17 @@ namespace NanoXLSX
         /// </summary>
         /// <remarks>
         /// The internal split height is based on the height of one or more rows. It also depends on various constants.<br/>
-        /// This method is derived from the Perl implementation by John McNamara (<a href="https://stackoverflow.com/a/5010899">https://stackoverflow.com/a/5010899</a>)
+        /// This method is derived from the Perl implementation by John McNamara (<a href="https://stackoverflow.com/a/5010899">https://stackoverflow.com/a/5010899</a>).<br/>
+        /// Negative row heights are automatically transformed to 0.
         /// </remarks>
         /// <param name="height">Target row(s) height (one or more rows, displayed in Excel)</param>
         /// <returns>The internal pane height, used in worksheet XML documents in case of worksheet splitting</returns>
         public static float GetInternalPaneSplitHeight(float height)
         {
+            if (height < 0)
+            {
+                height = 0f;
+            }
             return (float)Math.Floor(SPLIT_POINT_DIVIDER * height + SPLIT_HEIGHT_POINT_OFFSET);
         }
     }